requirements/common.txt
cssmin==0.2.0

Django==1.6.5
django-compressor==1.4
django-model-utils==2.2
django-suit==0.2.9

furl==0.3.95

requests==2.3.0

South==1.0
